chore: remove `hyperparameterEvolution` — feature complete, never enabled (Issue #3569)
Per-creature evolvable hyperparameters (Issue #1863) were fully implemented and
tested, but `hyperparameterEvolution.enabled` defaulted to `false` and no
consumer ever set it. `Mutator` never created a hyperparameter block, so no
creature ever carried a `hyperparameters` genome field, and both `Offspring`
branches — crossover and the disabled carry-through — were unreachable.

Removes the 13-field option surface (`NeatOptions` field, `CoerceNumeric`
mirror, both `keyof` unions, `NeatArguments` field, `NeatConfig` wiring,
`parseHyperparameterEvolution` and its re-export), `HyperparameterConfig.ts`,
`HyperparameterEvolution.ts`, the `hyperparameters` genome field with its
export/import/clone plumbing, the `docs/snapshot-schema.json` entry, the four
`mod.ts` public exports, and the `hyperparameterEvolution` lever from
`bench/EvolutionPaceLeverComparison.ts`.

`computeSpeciesDiversity` was the one live consumer of the removed module —
moved to `src/NEAT/SpeciesDiversity.ts` with its tests; adaptive population
sizing behaves identically.

Adds a regression guard asserting the parsed config no longer carries the key,
repins the `NeatArguments` top-level count at 109, and re-runs the pace-lever
bench so `docs/PERFORMANCE_RESEARCH.md` reports the surviving matrix honestly
(the four remaining levers are unchanged; `fast (combined)` moves 11 → 15
generations now that it is plateauDetection + mcmc only).

Existing creature JSON carrying a `hyperparameters` field still loads — the
field is ignored and dropped on the next export.

Vibe-Coder-Run-Id: vibe-msdxcmr2-7447e0
