+++ docs/API_REFERENCE.md
@@ -36,17 +36,17 @@ flowchart TD
 
 ## 📚 Detail docs
 
-| Topic                                                   | Covers                                                                                                                                                                                                                                                                                    |
-| ------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| **[Creature](api/CREATURE.md)**                         | `Creature` lifecycle (`activate`, `exportJSON`, `fromJSON`), the `CreatureFactory` helpers, `CRISPR` gene-editing, `CreatureUtil`, serialisation types (`CreatureExport`, `NeuronExport`, `SynapseExport`, `CreatureTrace`), `Upgrade` / `upgradeTwo`, `TypedTopology`.                   |
-| **[Configuration](api/CONFIGURATION.md)**               | `NeatOptions` / `NeatOptionsInput`, every sub-config (plateau, stability, regularisation, MCMC, output range, hyperparameter evolution, adaptive population, parallel evaluation, data fuzzing, selection pressure), training-event callbacks, `Logger`, and the `RandomNumberGenerator`. |
-| **[Costs & Activations](api/COSTS_AND_ACTIVATIONS.md)** | `Costs` registry, `CostInterface`, the built-in cost names, `costNameToTaskDescriptor` and the task-descriptor types, and the 39 activation (squash) functions.                                                                                                                           |
-| **[Evolution](api/EVOLUTION.md)**                       | `Creature.evolveDir()`, the reinforcement-learning `Creature.evolveRL()` + `EpisodeAdapter`, `Selection`, `Mutation`, configuration presets, plateau detection, speciation (`Species` / `Genus`), and the specialist pipeline.                                                            |
-| **[Training](api/TRAINING.md)**                         | `BackPropagationOptions`, `TrainOptions`, sparse-training controls, and the synthetic-synapse densification step.                                                                                                                                                                         |
-| **[Discovery](api/DISCOVERY.md)**                       | Discovery formatting utilities, `DiscoveryEvaluationSummary`, orphan-dir cleanup, and disk-space monitoring.                                                                                                                                                                              |
-| **[Interop & tooling](api/INTEROP.md)**                 | Transfer learning (checkpoints, population seeding), ONNX export, topology DOT/JSON export, and the Intelligent Design squash optimiser.                                                                                                                                                  |
-| **[Compute / WASM](api/COMPUTE.md)**                    | Worker WASM preloading, the activation LRU cache controls, and unified cache diagnostics.                                                                                                                                                                                                 |
-| **[Errors](api/ERRORS.md)**                             | `CrisprError`, `BreedExhaustionError`, `DatasetError`, and the `ValidationError` shape callers should `catch`.                                                                                                                                                                            |
+| Topic                                                   | Covers                                                                                                                                                                                                                                                                  |
+| ------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| **[Creature](api/CREATURE.md)**                         | `Creature` lifecycle (`activate`, `exportJSON`, `fromJSON`), the `CreatureFactory` helpers, `CRISPR` gene-editing, `CreatureUtil`, serialisation types (`CreatureExport`, `NeuronExport`, `SynapseExport`, `CreatureTrace`), `Upgrade` / `upgradeTwo`, `TypedTopology`. |
+| **[Configuration](api/CONFIGURATION.md)**               | `NeatOptions` / `NeatOptionsInput`, every sub-config (plateau, stability, regularisation, MCMC, output range, adaptive population, parallel evaluation, data fuzzing, selection pressure), training-event callbacks, `Logger`, and the `RandomNumberGenerator`.         |
+| **[Costs & Activations](api/COSTS_AND_ACTIVATIONS.md)** | `Costs` registry, `CostInterface`, the built-in cost names, `costNameToTaskDescriptor` and the task-descriptor types, and the 39 activation (squash) functions.                                                                                                         |
+| **[Evolution](api/EVOLUTION.md)**                       | `Creature.evolveDir()`, the reinforcement-learning `Creature.evolveRL()` + `EpisodeAdapter`, `Selection`, `Mutation`, configuration presets, plateau detection, speciation (`Species` / `Genus`), and the specialist pipeline.                                          |
+| **[Training](api/TRAINING.md)**                         | `BackPropagationOptions`, `TrainOptions`, sparse-training controls, and the synthetic-synapse densification step.                                                                                                                                                       |
+| **[Discovery](api/DISCOVERY.md)**                       | Discovery formatting utilities, `DiscoveryEvaluationSummary`, orphan-dir cleanup, and disk-space monitoring.                                                                                                                                                            |
+| **[Interop & tooling](api/INTEROP.md)**                 | Transfer learning (checkpoints, population seeding), ONNX export, topology DOT/JSON export, and the Intelligent Design squash optimiser.                                                                                                                                |
+| **[Compute / WASM](api/COMPUTE.md)**                    | Worker WASM preloading, the activation LRU cache controls, and unified cache diagnostics.                                                                                                                                                                               |
+| **[Errors](api/ERRORS.md)**                             | `CrisprError`, `BreedExhaustionError`, `DatasetError`, and the `ValidationError` shape callers should `catch`.                                                                                                                                                          |
 
 ## 💡 Minimal example
