+++ src/NEAT/AdaptivePopulationSizer.ts
@@ -6,7 +6,7 @@
  */
 
 import type { RequiredAdaptivePopulationConfig } from "@config/AdaptivePopulationConfig.ts";
-import { computeSpeciesDiversity } from "@neat/HyperparameterEvolution.ts";
+import { computeSpeciesDiversity } from "@neat/SpeciesDiversity.ts";
 
 /**
  * Computes the effective population size based on diversity and convergence.
