+++ src/breed/Breed.ts
@@ -183,7 +183,6 @@ export class Breed {
         geneticCompatibilityThreshold: config.geneticCompatibilityThreshold,
         interSpeciesCrossoverThreshold: config.interSpeciesCrossoverThreshold,
         forwardOnly: config.feedbackLoop !== true,
-        hyperparameterEvolution: config.hyperparameterEvolution,
       },
     );
