+++ src/breed/ParallelBreeding.ts
@@ -468,7 +468,6 @@ export class ParallelBreeding {
               geneticCompatibilityThreshold:
                 config.geneticCompatibilityThreshold,
               forwardOnly: config.feedbackLoop !== true,
-              hyperparameterEvolution: config.hyperparameterEvolution,
               subPhaseAccumulator: acc,
             },
           );
